ATLASRECTS-5201: Double to float rel22 FTAG

See merge request atlas/athena!36973

+++ Event/xAOD/xAODBTagging/xAODBTagging/ftagfloat_t.h
@@ -9,6 +9,6 @@
 /// https://its.cern.ch/jira/browse/ATLASRECTS-5201
 ///
 
-using ftagfloat_t = double;
+using ftagfloat_t = float;
 
 #endif
